% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score.R
\name{get_box_score}
\alias{get_box_score}
\title{Gets box score scoring summary data}
\usage{
get_box_score(..., progress = TRUE)
}
\arguments{
\item{...}{Function takes an object from \code{get_schedule()}. Function requires a \code{url}, \code{league}, and \code{season}. Additional data may be supplied. All of this information comes directly from \code{get_schedule()}.}

\item{progress}{Sets a Progress Bar. Defaults to \code{TRUE}.}
}
\description{
Returns a data frame of the box score scoring summary (including goal, primary assist, and secondary assist) for specified URL.
}
