% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Gets schedule data and box score URLs for specified league and season}
\usage{
get_schedule(league, season, ..., progress = TRUE)
}
\arguments{
\item{league}{Leagues from which the user wants to scrape data ("OHL" and/or "WHL" and/or "QMJHL").}

\item{season}{Seasons for which the user wants to scrape data. Must be of the form \code{2017-18}, \code{1995-96}, etc.}

\item{...}{Allows the user to supply other information to the function. If you don't know what this means, then don't worry about it.}

\item{progress}{Sets a Progress Bar. Defaults to \code{TRUE}.}
}
\description{
Returns a data frame of league schedule and game data for user supplied leagues & seasons.
}
\examples{
get_schedule("OHL", "2018-19")

get_schedule("QMJHL", "1999-00", progress = FALSE)

get_schedule(c("OHL", "WHL", "QMJHL"), c("2018-19", "2017-18"))

}
