% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score.R
\name{get_box_score}
\alias{get_box_score}
\title{Gets box score scoring summary data}
\usage{
get_box_score(..., progress = TRUE)
}
\arguments{
\item{...}{Function takes an object from \code{get_schedule()}. Function requires a \code{url}, \code{league}, and \code{season}. Additional data may be supplied. All of this information comes directly from \code{get_schedule()}.}

\item{progress}{Sets a Progress Bar. Defaults to \code{TRUE}.}
}
\description{
Returns a data frame of the box score scoring summary (including goal, primary assist, and secondary assist) for specified URL.
}
\examples{
## Use get_schedule() to get the games and URLs desired
games <- get_schedule("OHL", "2018-19")

## Easy peasy get box score scoring summary data
get_box_score(games)

## If you want to see 1 game, try using dplyr::slice to only select the 1st game
games \%>\% 
  slice(1) \%>\% 
  get_box_score()

}
